Update README.md
Reorged by current table of contents of the book. Added 6 ("The package within") and 9 ("Licensing").

I'm going to do a quick check of learning objectives before session 2 at least so we can make sure we don't gloss over a pre-req, and to otherwise weigh in with what I think about the order... but for now at least everything is included!